feat/add validations config and some open api tags

+++ src/modules/customer/customer.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CustomerDTO } from '../../types/dtos/customer/customer.dto';
 import { CustomerUpdateDTO } from '../../types/dtos/customer/customerUpdate.dto';
 import { Customer } from '@prisma/client';
@@ -36,13 +36,37 @@ export class CustomerService {
     });
   }
 
-  delete(id: string): Promise<Customer> {
+  async delete(id: string): Promise<Customer> {
+    const customer = this.prisma.customer.findFirst({
+      where: { id },
+    });
+    if (!customer) {
+      throw new HttpException(
+        {
+          error: 'Id de cliente não encontrado',
+        },
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return this.prisma.customer.delete({
       where: { id },
     });
   }
 
-  update(id: string, customerData: CustomerUpdateDTO): Promise<Customer> {
+  async update(id: string, customerData: CustomerUpdateDTO): Promise<Customer> {
+    const customer = await this.prisma.customer.findFirst({
+      where: { id },
+    });
+    if (!customer) {
+      throw new HttpException(
+        {
+          error: 'Id de cliente não encontrado',
+        },
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return this.prisma.customer.update({
       data: customerData,
       where: { id },
