+++ src/modules/schedule/schedule.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { Schedule } from '@prisma/client';
 import DateHelper from 'src/helpers/dateHelper';
 import { PrismaService } from 'src/modules/prisma/prisma.service';
@@ -8,7 +8,7 @@ import { ScheduleUpdateDTO } from '../../types/dtos/schedule/scheduleUpdate.dto'
 @Injectable()
 export class ScheduleService {
   constructor(private readonly prisma: PrismaService) {}
-  async getOne(id: string): Promise<Schedule> {
+  getOne(id: string): Promise<Schedule> {
     return this.prisma.schedule.findFirst({
       where: { id },
       include: {
@@ -18,7 +18,7 @@ export class ScheduleService {
     });
   }
 
-  async getAllByDate(date: string, type: string): Promise<Schedule[]> {
+  getAllByDate(date: string, type: string): Promise<Schedule[]> {
     const queryDate = DateHelper.getParamDate(date, type);
     return this.prisma.schedule.findMany({
       where: {
@@ -34,19 +34,43 @@ export class ScheduleService {
     });
   }
 
-  async create(scheduleData: ScheduleDTO): Promise<Schedule> {
+  create(scheduleData: ScheduleDTO): Promise<Schedule> {
     return this.prisma.schedule.create({
       data: scheduleData,
     });
   }
 
   async delete(id: string): Promise<Schedule> {
+    const schedule = await this.prisma.schedule.findFirst({
+      where: { id },
+    });
+    if (!schedule) {
+      throw new HttpException(
+        {
+          error: 'Id de horário não encontrado',
+        },
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return this.prisma.schedule.delete({
       where: { id },
     });
   }
 
   async update(id: string, scheduleData: ScheduleUpdateDTO): Promise<Schedule> {
+    const schedule = await this.prisma.schedule.findFirst({
+      where: { id },
+    });
+    if (!schedule) {
+      throw new HttpException(
+        {
+          error: 'Id de horário não encontrado',
+        },
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return this.prisma.schedule.update({
       data: scheduleData,
       where: { id },
