+++ src/modules/type/type.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { Type } from '@prisma/client';
 import { PrismaService } from 'src/modules/prisma/prisma.service';
 import { TypeDTO } from '../../types/dtos/type/type.dto';
@@ -9,27 +9,51 @@ export class TypeService {
   constructor(private readonly prisma: PrismaService) {}
   async getOne(id: string): Promise<Type> {
     return this.prisma.type.findFirst({
-      where: { id }
+      where: { id },
     });
   }
 
-  async getAll(): Promise<Type[]> {
+  getAll(): Promise<Type[]> {
     return this.prisma.type.findMany();
   }
 
-  async create(typeData: TypeDTO): Promise<Type> {
+  create(typeData: TypeDTO): Promise<Type> {
     return this.prisma.type.create({
       data: typeData,
     });
   }
 
   async delete(id: string): Promise<Type> {
+    const type = await this.prisma.type.findFirst({
+      where: { id },
+    });
+    if (!type) {
+      throw new HttpException(
+        {
+          error: 'Id de tipo não encontrado',
+        },
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return this.prisma.type.delete({
       where: { id },
     });
   }
 
   async update(id: string, typeData: TypeUpdateDTO): Promise<Type> {
+    const customer = await this.prisma.customer.findFirst({
+      where: { id },
+    });
+    if (!customer) {
+      throw new HttpException(
+        {
+          error: 'Id de cliente não encontrado',
+        },
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return this.prisma.type.update({
       data: typeData,
       where: { id },
